// src/app/model/response.model.ts
export interface Response {
  lastName: string;
  firstName: string;
  contactNumber: string;
  age: number;
  surveyDate: Date;
  pizza: string;
  pasta: string;
  papAndWors: string;
  chickenStirFry: string;
  beefStirFry: string;
  other: string;
  eatOut: number;
  watchMovie: number;
  watchTv: number;
  listenToRadio: number;
}